Add Miner & Delete BaseMiner
